+++ resources/views/directives/print/overview.blade.php
@@ -29,7 +29,7 @@
 </head>
 <body>
 
-<div class="no-print"><button onclick="window.print()">Drucken</button></div>
+<div class="no-print"><button type="button" data-print>Drucken</button><script src="/js/print-button.js"></script></div>
 
 <div class="head">
     <div class="title">Übersicht LTA / TM</div>